fix(js): rename magma.js back to script-evaluator.js
The old forms load it from the server and expect it to be called script-evaluator.js.

Fixes #9135
