+++ molgenis-js/src/main/java/org/molgenis/js/magma/JsMagmaScriptContext.java
@@ -31,7 +31,8 @@ public class JsMagmaScriptContext {
   private final Context context;
 
   static {
-    SOURCES = Stream.of("/js/magma.js").map(JsMagmaScriptContext::getSource).collect(toList());
+    SOURCES =
+        Stream.of("/js/script-evaluator.js").map(JsMagmaScriptContext::getSource).collect(toList());
   }
 
   private static Source getSource(String resourceName) {
